Fixed windows compilation.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@34337 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/bdb/bdb_blobcache.cpp
@@ -2882,8 +2882,8 @@ purge_start:
                             // we need this in order not to move point of restart
                             // too agressively
 
-                            double delta = fabs(m_NextExpTime - exp_time);
-                            double fraction = fabs(delta / m_NextExpTime);
+                            double delta = fabs(double(m_NextExpTime - exp_time));
+                            double fraction = fabs(double(delta / m_NextExpTime));
                             if (fraction > 0.1 && delta > m_PurgeThreadDelay) {
                                 next_exp_key = key;
                                 m_NextExpTime = exp_time;
